use test_common in farm
