+++ x/farm/keeper/test_common.go
@@ -3,6 +3,7 @@ package keeper
 import (
 	"bytes"
 	"fmt"
+	swap "github.com/okex/okexchain/x/ammswap"
 	"strconv"
 	"testing"
 
@@ -12,7 +13,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/bank"
 	"github.com/cosmos/cosmos-sdk/x/supply"
-	swap "github.com/okex/okexchain/x/ammswap/keeper"
 	swaptypes "github.com/okex/okexchain/x/ammswap/types"
 	"github.com/okex/okexchain/x/farm/types"
 	"github.com/okex/okexchain/x/params"
@@ -38,20 +38,43 @@ type MockFarmKeeper struct {
 	SupplyKeeper supply.Keeper
 	MountedStore store.MultiStore
 	AccKeeper    auth.AccountKeeper
+	BankKeeper   bank.Keeper
+	TokenKeeper  token.Keeper
+	SwapKeeper   swap.Keeper
 }
 
-func NewMockFarmKeeper(k Keeper, keyStoreKey sdk.StoreKey, sKeeper supply.Keeper,
-	ms store.MultiStore, accKeeper auth.AccountKeeper) MockFarmKeeper {
+func NewMockFarmKeeper(
+	k Keeper, keyStoreKey sdk.StoreKey, sKeeper supply.Keeper,
+	ms store.MultiStore, accKeeper auth.AccountKeeper, bankKeeper bank.Keeper,
+	tokenKeeper token.Keeper, swapKeeper swap.Keeper,
+) MockFarmKeeper {
 	return MockFarmKeeper{
 		k,
 		keyStoreKey,
 		sKeeper,
 		ms,
 		accKeeper,
+		bankKeeper,
+		tokenKeeper,
+		swapKeeper,
 	}
 }
 
-func getKeeper(t *testing.T) (sdk.Context, MockFarmKeeper) {
+func TestBeginBlocker(t *testing.T) {
+	//ctx, mk := GetKeeper(t)
+	//k := mk.Keeper
+	//
+	//// TODO issue token
+	//// TODO create swap
+	//// TODO test farm
+	//
+	//for i := int64(1); i < 10; i++ {
+	//	ctx = ctx.WithBlockHeight(i)
+	//	BeginBlocker(ctx, abci.RequestBeginBlock{Header: abci.Header{Height: i}}, k)
+	//}
+}
+
+func GetKeeper(t *testing.T) (sdk.Context, MockFarmKeeper) {
 
 	// 0.1 init store key
 	keyFarm := sdk.NewKVStoreKey(types.StoreKey)
@@ -107,15 +130,14 @@ func getKeeper(t *testing.T) (sdk.Context, MockFarmKeeper) {
 
 	// 1.3 init bank keeper
 	feeCollectorAcc := supply.NewEmptyModuleAccount(auth.FeeCollectorName)
-	farmAcc := supply.NewEmptyModuleAccount(types.ModuleName)
-	yieldAcc := supply.NewEmptyModuleAccount(types.YieldFarmingAccount)
-	mintAcc := supply.NewEmptyModuleAccount(types.MintFarmingAccount)
+	farmAcc := supply.NewEmptyModuleAccount(types.ModuleName, supply.Burner, supply.Minter)
+	yieldFarmingAccount := supply.NewEmptyModuleAccount(types.YieldFarmingAccount, supply.Burner, supply.Minter)
+	mintFarmingAccount := supply.NewEmptyModuleAccount(types.MintFarmingAccount, supply.Burner, supply.Minter)
+	swapModuleAccount := supply.NewEmptyModuleAccount(swap.ModuleName, supply.Burner, supply.Minter)
 
 	blacklistedAddrs := make(map[string]bool)
 	blacklistedAddrs[feeCollectorAcc.String()] = true
 	blacklistedAddrs[farmAcc.String()] = true
-	blacklistedAddrs[yieldAcc.String()] = true
-	blacklistedAddrs[mintAcc.String()] = true
 
 	bk := bank.NewBaseKeeper(ak, pk.Subspace(bank.DefaultParamspace), bank.DefaultCodespace, blacklistedAddrs)
 	// fill all the addresses with some coins
@@ -129,17 +151,19 @@ func getKeeper(t *testing.T) (sdk.Context, MockFarmKeeper) {
 
 	// 1.4 init supply keeper
 	maccPerms := map[string][]string{
-		auth.FeeCollectorName: nil,
-		types.ModuleName:      nil,
-		types.MintFarmingAccount: nil,
+		auth.FeeCollectorName:     nil,
+		types.ModuleName:          nil,
 		types.YieldFarmingAccount: nil,
+		types.MintFarmingAccount:  nil,
+		swap.ModuleName:           {supply.Burner, supply.Minter},
 	}
 	sk := supply.NewKeeper(cdc, keySupply, ak, bk, maccPerms)
 	sk.SetSupply(ctx, supply.NewSupply(sdk.NewDecCoinsFromDec(sdk.DefaultBondDenom, sdk.NewDec(1000000000))))
 	sk.SetModuleAccount(ctx, feeCollectorAcc)
 	sk.SetModuleAccount(ctx, farmAcc)
-	sk.SetModuleAccount(ctx, yieldAcc)
-	sk.SetModuleAccount(ctx, mintAcc)
+	sk.SetModuleAccount(ctx, yieldFarmingAccount)
+	sk.SetModuleAccount(ctx, mintFarmingAccount)
+	sk.SetModuleAccount(ctx, swapModuleAccount)
 
 	// 1.5 init token keeper
 	tk := token.NewKeeper(bk, pk.Subspace(token.DefaultParamspace), auth.FeeCollectorName, sk, keyToken, keyLock, cdc, false)
@@ -151,7 +175,7 @@ func getKeeper(t *testing.T) (sdk.Context, MockFarmKeeper) {
 	fk := NewKeeper(auth.FeeCollectorName, sk, tk, swapKeeper, pk.Subspace(types.DefaultParamspace), keyFarm, cdc)
 	fk.SetParams(ctx, types.DefaultParams())
 	// 2. init mock keeper
-	mk := NewMockFarmKeeper(fk, keyFarm, sk, ms, ak)
+	mk := NewMockFarmKeeper(fk, keyFarm, sk, ms, ak, bk, tk, swapKeeper)
 
 	//// 3. init mockApp
 	//mApp := mock.NewApp()
